+++ .env
@@ -1,1 +1,2 @@
-EXPO_PUBLIC_CLERK_PUBLISHABLE_KEY=pk_test_ZnJhbmstbWFsYW11dGUtMTUuY2xlcmsuYWNjb3VudHMuZGV2JA
+EXPO_PUBLIC_CLERK_PUBLISHABLE_KEY=pk_test_Zmx5aW5nLXZpcGVyLTMyLmNsZXJrLmFjY291bnRzLmRldiQ
+DATABASE_URL=postgresql://JSM_Uber_owner:CrP4G8IiLylM@ep-patient-math-a5e5k88w.us-east-2.aws.neon.tech/JSM_Uber?sslmode=require
