+++ app/(auth)/_layout.tsx
@@ -1,17 +1,15 @@
-import { Stack } from 'expo-router';
-import React from 'react';
+import { Stack } from "expo-router";
+import React from "react";
 
-
-const Layout = () =>{
-
+const Layout = () => {
   return (
-      <Stack>
-        <Stack.Screen name="log-in" options={{ headerShown: false }} />
-        <Stack.Screen name="sign-up" options={{ headerShown: false }} />                   
-        <Stack.Screen name="guides" options={{ headerShown: false }} />                    
-        <Stack.Screen name="guide-Buying" options={{ headerShown: false }} />
-        <Stack.Screen name="forgot-password" options={{headerShown: false}}/>
-      </Stack>
+    <Stack>
+      <Stack.Screen name="log-in" options={{ headerShown: false }} />
+      <Stack.Screen name="sign-up" options={{ headerShown: false }} />
+      {/* <Stack.Screen name="guides" options={{ headerShown: false }} />                    
+        <Stack.Screen name="guide-Buying" options={{ headerShown: false }} /> */}
+      <Stack.Screen name="forgot-password" options={{ headerShown: false }} />
+    </Stack>
   );
-}
+};
 export default Layout;
