+++ app/(auth)/sign-up.tsx
@@ -1,7 +1,7 @@
 import { useSignUp } from "@clerk/clerk-expo";
 import { Link, router } from "expo-router";
 import { useState } from "react";
-import { Alert, Image, ScrollView, Text, View, TouchableOpacity } from "react-native";
+import { Alert, Image, ScrollView, Text, View } from "react-native";
 import { ReactNativeModal } from "react-native-modal";
 
 import CustomButton from "@/components/CustomButton";
@@ -9,10 +9,12 @@ import InputField from "@/components/InputField";
 import OAuth from "@/components/OAuth";
 import { icons, images } from "@/constants";
 import React from "react";
+import { fetchAPI } from "@/lib/fetch";
 
 const SignUp = () => {
   const { isLoaded, signUp, setActive } = useSignUp();
   const [showSuccessModal, setShowSuccessModal] = useState(false);
+
   const [form, setForm] = useState({
     name: "",
     email: "",
@@ -23,7 +25,6 @@ const SignUp = () => {
     error: "",
     code: "",
   });
-  const [isPasswordVisible, setIsPasswordVisible] = useState(false); // State for password visibility
 
   const onSignUpPress = async () => {
     if (!isLoaded) return;
@@ -38,11 +39,12 @@ const SignUp = () => {
         state: "pending",
       });
     } catch (err: any) {
+      // See https://clerk.com/docs/custom-flows/error-handling
+      // for more info on error handling
       console.log(JSON.stringify(err, null, 2));
       Alert.alert("Error", err.errors[0].longMessage);
     }
   };
-
   const onPressVerify = async () => {
     if (!isLoaded) return;
     try {
@@ -50,11 +52,20 @@ const SignUp = () => {
         code: verification.code,
       });
       if (completeSignUp.status === "complete") {
-        await setActive({ session: completeSignUp.createdSessionId });
-        setVerification({
-          ...verification,
-          state: "success",
-        });
+        await fetchAPI(
+          "https://b096-2605-8d80-6c3-f7e8-c4f-9f87-302f-3eed.ngrok-free.app/api/user",
+          {
+            method: "POST",
+            headers: {
+              "Content-Type": "application/json",
+            },
+            body: JSON.stringify({
+              name: form.name,
+              email: form.email,
+              clerkId: completeSignUp.createdUserId,
+            }),
+          }
+        );
       } else {
         setVerification({
           ...verification,
@@ -63,6 +74,8 @@ const SignUp = () => {
         });
       }
     } catch (err: any) {
+      // See https://clerk.com/docs/custom-flows/error-handling
+      // for more info on error handling
       setVerification({
         ...verification,
         error: err.errors[0].longMessage,
@@ -70,7 +83,6 @@ const SignUp = () => {
       });
     }
   };
-
   return (
     <ScrollView className="flex-1 bg-white">
       <View className="flex-1 bg-white">
@@ -96,34 +108,25 @@ const SignUp = () => {
             value={form.email}
             onChangeText={(value) => setForm({ ...form, email: value })}
           />
-          <View className="relative">
-            <InputField
-              label="Password"
-              placeholder="Enter password"
-              icon={icons.lock}
-              secureTextEntry={!isPasswordVisible} // Use the state to determine visibility
-              textContentType="password"
-              value={form.password}
-              onChangeText={(value) => setForm({ ...form, password: value })}
-            />
-            <TouchableOpacity
-              onPress={() => setIsPasswordVisible(!isPasswordVisible)}
-              style={{
-                position: 'absolute',
-                right: 10,
-                top: 15,
-                zIndex: 1, // Ensure it is above other elements
-              }}
-            >
-              <Image
-                source={isPasswordVisible ? icons.eyeOpen : icons.eyeClose} // Change icon based on visibility
-                style={{ width: 24, height: 24 }}
-              />
-            </TouchableOpacity>
-          </View>
-          <CustomButton title="Sign Up" onPress={onSignUpPress} className="mt-6" />
+          <InputField
+            label="Password"
+            placeholder="Enter password"
+            icon={icons.lock}
+            secureTextEntry={true}
+            textContentType="password"
+            value={form.password}
+            onChangeText={(value) => setForm({ ...form, password: value })}
+          />
+          <CustomButton
+            title="Sign Up"
+            onPress={onSignUpPress}
+            className="mt-6"
+          />
           <OAuth />
-          <Link href="/log-in" className="text-lg text-center text-general-200 mt-5">
+          <Link
+            href="/log-in"
+            className="text-lg text-center text-general-200 mt-5"
+          >
             Already have an account?{" "}
             <Text className="text-primary-500">Log In</Text>
           </Link>
@@ -137,7 +140,9 @@ const SignUp = () => {
           }}
         >
           <View className="bg-white px-7 py-9 rounded-2xl min-h-[300px]">
-            <Text className="font-JakartaExtraBold text-2xl mb-2">Verification</Text>
+            <Text className="font-JakartaExtraBold text-2xl mb-2">
+              Verification
+            </Text>
             <Text className="font-Jakarta mb-5">
               We've sent a verification code to {form.email}.
             </Text>
@@ -147,18 +152,31 @@ const SignUp = () => {
               placeholder={"12345"}
               value={verification.code}
               keyboardType="numeric"
-              onChangeText={(code) => setVerification({ ...verification, code })}
+              onChangeText={(code) =>
+                setVerification({ ...verification, code })
+              }
             />
             {verification.error && (
-              <Text className="text-red-500 text-sm mt-1">{verification.error}</Text>
+              <Text className="text-red-500 text-sm mt-1">
+                {verification.error}
+              </Text>
             )}
-            <CustomButton title="Verify Email" onPress={onPressVerify} className="mt-5 bg-success-500" />
+            <CustomButton
+              title="Verify Email"
+              onPress={onPressVerify}
+              className="mt-5 bg-success-500"
+            />
           </View>
         </ReactNativeModal>
         <ReactNativeModal isVisible={showSuccessModal}>
           <View className="bg-white px-7 py-9 rounded-2xl min-h-[300px]">
-            <Image source={images.check} className="w-[110px] h-[110px] mx-auto my-5" />
-            <Text className="text-3xl font-JakartaBold text-center">Verified</Text>
+            <Image
+              source={images.check}
+              className="w-[110px] h-[110px] mx-auto my-5"
+            />
+            <Text className="text-3xl font-JakartaBold text-center">
+              Verified
+            </Text>
             <Text className="text-base text-gray-400 font-Jakarta text-center mt-2">
               You have successfully verified your account.
             </Text>
@@ -166,7 +184,7 @@ const SignUp = () => {
               title="Browse Home"
               onPress={() => {
                 setShowSuccessModal(false);
-                router.push('/(root)/Homepage');
+                router.push("../Homepage");
               }}
               className="mt-5"
             />
@@ -176,5 +194,4 @@ const SignUp = () => {
     </ScrollView>
   );
 };
-
 export default SignUp;
