+++ app/(guides)/SellingGuides/StepOnePreSalePurchase.tsx
@@ -1,0 +1,261 @@
+import React, { useRef, useState, useEffect, forwardRef } from "react";
+import {
+  View,
+  Text,
+  TouchableOpacity,
+  ScrollView,
+  Image,
+  FlatList,
+  Linking,
+  Animated,
+} from "react-native";
+import { useRouter } from "expo-router";
+import { images } from "@/constants";
+import { SafeAreaView } from "react-native-safe-area-context";
+import { Question } from "@/types/type";
+
+// Define subStepsOne outside of the component
+const subStepsOneSelling = [
+  {
+    title: "a) Review Your Income",
+    bullets: [
+      {
+        text: "Create a Budget",
+        detail:
+          "Establish a clear monthly budget to track your income and expenses.",
+        resource: "https://www.mint.com",
+      },
+      {
+        text: "Calculate Debt-to-Income Ratio",
+        detail:
+          "Analyze your total monthly debt payments compared to your gross monthly income.",
+        resource:
+          "https://www.calculator.net/debt-to-income-ratio-calculator.html",
+      },
+      {
+        text: "Identify Other Income Sources",
+        detail:
+          "Include any additional income sources such as bonuses, freelance work, or rental income.",
+      },
+    ],
+  },
+];
+
+const StepOnePreSalePurchase = () => {
+  const [expanded, setExpanded] = useState(false); // State to track the expanded state of the main step
+  const [showQuestion, setShowQuestion] = useState(false);
+  const [selectedOption, setSelectedOption] = useState<string | null>(null); // Define type for selectedOption
+  const [isCorrect, setIsCorrect] = useState<boolean | null>(null);
+  const [explanation, setExplanation] = useState<string | null>(null);
+  const [fadeAnim] = useState(new Animated.Value(0)); // Animation value for fade
+  const [currentQuestionIndex, setCurrentQuestionIndex] = useState(0); // Track the current question index
+
+  // State to track the expanded state of each substep
+  const [expandedSubSteps, setExpandedSubSteps] = useState<boolean[]>(
+    Array(subStepsOneSelling.length).fill(false)
+  );
+
+  const questionsStepOne: Question[] = [
+    {
+      text: "What is considered a good credit score for a better mortgage rate?",
+      options: ["A- Below 600", "B- 600-699", "C- 750 above"],
+      correctAnswer: "C- 750 above",
+      correctExplanation:
+        "That's right! A credit score of 750 or higher generally qualifies for the best mortgage rates. If your score is lower, you might consider steps to improve it.",
+      incorrectExplanation:
+        "Not quite. A score of 750 or above is ideal for getting the best mortgage rates. It’s worth checking your credit score and improving it before you apply for a mortgage.",
+    },
+    {
+      text: "What is the typical down payment required for a home?",
+      options: [
+        "A) 1% of the home’s price",
+        "B) 5% to 20% of the home’s price",
+        "C) 25% of the home’s price",
+        "D) 50% of the home’s price",
+      ],
+      correctAnswer: "B) 5% to 20% of the home’s price",
+      correctExplanation:
+        "Great! The typical down payment ranges from 5% to 20% of the home’s price, depending on the market and your financial situation.",
+      incorrectExplanation:
+        "Not quite. A down payment usually falls between 5% and 20%, though some buyers might be eligible for programs with lower requirements.",
+    },
+  ];
+
+  const handleExpand = () => {
+    setExpanded(!expanded);
+  };
+
+  const handleSubStepExpand = (index: number) => {
+    const newExpandedState = [...expandedSubSteps];
+    newExpandedState[index] = !newExpandedState[index]; // Toggle the specific substep
+    setExpandedSubSteps(newExpandedState);
+  };
+
+  const handleOptionSelect = (option: string, question: Question) => {
+    setSelectedOption(option);
+    const correct = option === question.correctAnswer;
+    setIsCorrect(correct);
+    setExplanation(
+      correct ? question.correctExplanation : question.incorrectExplanation
+    );
+
+    // Fade in explanation
+    Animated.timing(fadeAnim, {
+      toValue: 1,
+      duration: 500,
+      useNativeDriver: true,
+    }).start(() => {
+      setTimeout(() => {
+        // Fade out after 3 seconds
+        Animated.timing(fadeAnim, {
+          toValue: 0,
+          duration: 500,
+          useNativeDriver: true,
+        }).start(() => {
+          setExplanation(null);
+          //setIsCorrect(null); // Reset highlighting after fading out
+          //setSelectedOption(null); // Reset selected option
+        });
+      }, 3000);
+    });
+  };
+
+  const nextQuestion = () => {
+    // Check if this is the last question
+    if (currentQuestionIndex < questionsStepOne.length - 1) {
+      setCurrentQuestionIndex(currentQuestionIndex + 1);
+    } else {
+      setCurrentQuestionIndex(0); // Reset to the first question
+    }
+    setSelectedOption(null); // Reset selected option for the next question
+    setIsCorrect(null); // Reset correctness for the next question
+  };
+
+  const nextStep = () => {
+    // Logic for proceeding to the next step can be added here
+    console.log("Proceeding to the next step...");
+  };
+
+  return (
+    // <ScrollView ref={scrollViewRef}>
+    <View className="flex-1 bg-white p-4">
+      <TouchableOpacity
+        onPress={handleExpand}
+        className="bg-blue-100 p-4 rounded-lg flex-row items-center justify-center mb-2"
+      >
+        <Text className="text-xl font-bold text-blue-700">
+          1. Pre-Sale-Purchase
+        </Text>
+      </TouchableOpacity>
+
+      {expanded && (
+        <View className="bg-gray-200 p-4 rounded-lg">
+          <FlatList
+            data={subStepsOneSelling}
+            keyExtractor={(item) => item.title}
+            renderItem={({ item, index }) => (
+              <View className="mb-4">
+                <TouchableOpacity
+                  onPress={() => handleSubStepExpand(index)}
+                  className="flex-row items-center justify-between bg-gray-300 p-3 rounded-lg"
+                >
+                  <Text className="text-lg font-bold">{item.title}</Text>
+                  <Text>{expandedSubSteps[index] ? "-" : "+"}</Text>
+                </TouchableOpacity>
+                {expandedSubSteps[index] && (
+                  <View className="ml-4">
+                    {item.bullets.map((bullet, bulletIndex) => (
+                      <View key={bulletIndex} className="mb-2">
+                        <Text className="font-semibold">• {bullet.text}</Text>
+                        {bullet.detail && (
+                          <Text className="mt-1 text-sm text-gray-700">
+                            {bullet.detail}
+                          </Text>
+                        )}
+                        {bullet.resource && (
+                          <Text className="mt-1 text-sm">
+                            <Text style={{ color: "black" }}>
+                              Helpful link:{" "}
+                            </Text>
+                            <Text
+                              className="text-blue-500 underline"
+                              onPress={() => Linking.openURL(bullet.resource)}
+                            >
+                              {bullet.resource}
+                            </Text>
+                          </Text>
+                        )}
+                      </View>
+                    ))}
+                  </View>
+                )}
+              </View>
+            )}
+          />
+        </View>
+      )}
+
+      {!showQuestion && expanded && (
+        <TouchableOpacity
+          onPress={() => setShowQuestion(true)}
+          className="mt-4 bg-blue-600 p-2 rounded-lg"
+        >
+          <Text className="text-white text-center">Test Your Knowledge</Text>
+        </TouchableOpacity>
+      )}
+
+      {showQuestion && (
+        <View className="mt-4">
+          <Text className="font-bold text-lg">
+            {questionsStepOne[currentQuestionIndex].text}
+          </Text>
+          {questionsStepOne[currentQuestionIndex].options.map((option) => (
+            <TouchableOpacity
+              key={option}
+              onPress={() =>
+                handleOptionSelect(
+                  option,
+                  questionsStepOne[currentQuestionIndex]
+                )
+              }
+              className={`p-2 rounded-lg mt-2 ${
+                selectedOption === option
+                  ? isCorrect
+                    ? "bg-green-200"
+                    : "bg-red-200"
+                  : "bg-gray-300"
+              }`}
+            >
+              <Text>{option}</Text>
+            </TouchableOpacity>
+          ))}
+          {explanation && (
+            <Animated.View style={{ opacity: fadeAnim }}>
+              <Text className="mt-2">{explanation}</Text>
+            </Animated.View>
+          )}
+        </View>
+      )}
+
+      {selectedOption && (
+        <View className="flex-row justify-between mt-4">
+          <TouchableOpacity
+            onPress={nextQuestion}
+            className="bg-blue-600 p-2 rounded-lg"
+          >
+            <Text className="text-white text-center">Next Question</Text>
+          </TouchableOpacity>
+          <TouchableOpacity
+            onPress={nextStep}
+            className="bg-green-600 p-2 rounded-lg"
+          >
+            <Text className="text-white text-center">Next Step</Text>
+          </TouchableOpacity>
+        </View>
+      )}
+    </View>
+    //</ScrollView>
+  );
+};
+
+export default StepOnePreSalePurchase;
