--- app/(root)/(tabs)/user.tsx
@@ -1,24 +1,0 @@
-import { SignedIn, SignedOut, useUser } from '@clerk/clerk-expo'
-import { Link } from 'expo-router'
-import React from 'react'
-import { Text, View } from 'react-native'
-
-export default function Page() {
-  const { user } = useUser()
-
-  return (
-    <View>
-      <SignedIn>
-        <Text>Hello {user?.emailAddresses[0].emailAddress}</Text>
-      </SignedIn>
-      <SignedOut>
-        <Link href="/log-in">
-          <Text>Sign In</Text>
-        </Link>
-        <Link href="/sign-up">
-          <Text>Sign Up</Text>
-        </Link>
-      </SignedOut>
-    </View>
-  )
-}
