--- app/(root)/log-in.tsx
@@ -1,6 +1,0 @@
-
-function LogInn(){      
-      return (
-        "hi"
-      ) 
-}
