+++ app/+not-found.tsx
@@ -1,14 +1,14 @@
-import { Link, Stack } from 'expo-router';
-import { StyleSheet, View, Text } from 'react-native';
-import React from 'react';
+import { Link, Stack } from "expo-router";
+import { StyleSheet, View, Text } from "react-native";
+import React from "react";
 
 export default function NotFoundScreen() {
   return (
     <>
-      <Stack.Screen options={{ title: 'Oops!' }} />
+      <Stack.Screen options={{ title: "Oops!" }} />
       <View style={styles.container}>
         <Text>This screen doesn't exist.</Text>
-        <Link href="/" style={styles.link}>
+        <Link href="/(root)/(tabs)/home" style={styles.link}>
           <Text>Go to home screen!</Text>
         </Link>
       </View>
@@ -19,8 +19,8 @@ export default function NotFoundScreen() {
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
+    alignItems: "center",
+    justifyContent: "center",
     padding: 20,
   },
   link: {
