+++ constants/index.ts
@@ -27,13 +27,13 @@ import check from "@/assets/images/check.png";
 import message from "@/assets/images/message.png";
 import noResult from "@/assets/images/no-result.png";
 import houseHeader from "@/assets/images/house-header.png";
-import buyingGuidesSelection from "@/assets/images/buying-guide-selection.png";
 import appLogo from "@/assets/images/app-logo.png";
 import primary from "@/assets/images/primary.png";
 import filters from "@/assets/icons/filters.png";import eyeOpen from "@/assets/icons/open-eye.png"
 import eyeClose from "@/assets/icons/close-eye.png"
 import deleteCan from "@/assets/icons/delete.png"
 import favIcon from "@/assets/icons/heart.png"
+import buyingGuidesSelection from "@/assets/images/buying-guide-selection.png";
 import sellingGuideSelection from "@/assets/images/Selling-Guide-Section.png";
 import rentingTenantGuideSelection from "@/assets/images/Tenant-Guide-Section.png";
 import rentingLandlordGuideSelection from "@/assets/images/Landlord-Guide-Section.png";
