+++ package.json
@@ -19,19 +19,24 @@
     "@expo/vector-icons": "^14.0.2",
     "@react-navigation/native": "^6.1.18",
     "@react-navigation/stack": "^6.4.1",
-    "expo": "~51.0.37",
+    "axios": "^1.7.8",
+    "cors": "^2.8.5",
+    "dotenv": "^16.4.5",
+    "expo": "~51.0.39",
     "expo-constants": "~16.0.2",
     "expo-font": "~12.0.9",
     "expo-image": "~1.13.0",
     "expo-linking": "~6.3.1",
     "expo-local-authentication": "^14.0.1",
-    "expo-router": "~3.5.23",
+    "expo-router": "~3.5.24",
     "expo-secure-store": "^13.0.2",
-    "expo-splash-screen": "~0.27.5",
+    "expo-splash-screen": "~0.27.7",
     "expo-status-bar": "~1.12.1",
     "expo-system-ui": "~3.0.7",
     "expo-web-browser": "~13.0.3",
+    "express": "^4.21.1",
     "nativewind": "^2.0.11",
+    "pg": "^8.13.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-native": "0.74.5",
